feat: fix(google-calendar): cast sync_days_ahead to int to prevent Ca… (#64)
…rbon type error

+++ app-modules/integration-google-calendar/config/google-calendar.php
@@ -2,5 +2,5 @@
 
 return [
     'service_account_credentials' => env('GOOGLE_SERVICE_ACCOUNT_CREDENTIALS'),
-    'sync_days_ahead' => env('GOOGLE_CALENDAR_SYNC_DAYS_AHEAD', 60),
+    'sync_days_ahead' => (int) env('GOOGLE_CALENDAR_SYNC_DAYS_AHEAD', 60),
 ];
